+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unpublished
 
-- Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the @masterportal/masterportalapi default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
+- Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
 - Feature: Add zoomLevel as plugin-agnostic map information to store.
 
 ## 1.1.0