+++ packages/core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
 - Feature: Add zoomLevel as plugin-agnostic map information to store.
-- Chore: Adding README information about listening to map client state.
+- Chore: Add README information about listening to map client state and getters.
 
 ## 1.1.0