+++ packages/core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Feature: Add hovered and selected feature to vuex store. This is an optional functionality that has to be explicitly enabled and works with the `@masterportal/masterportalapi` default marker design. See configuration parameter `extendedMasterportalapiMarkers`.
 - Feature: Add zoomLevel as plugin-agnostic map information to store.
-- Feature: Change the `background-color` of all `v-tooltip`s to `primary` and its `border` to `primaryContrast` to be more visible.
+- Feature: Change the `background-color` of all `v-tooltip`s to `#595959` and its `border` to `#fff` to be more visible. It now always has a contrast of 7, which is quite enough for AAA of WCAG.
 - Feature: Add new state variable `hasSmallDisplay` which is updated on `resize` of the `window`.
 
 ## 1.1.0